\name{plot_spec}
\alias{plot_spec}
\title{Output plot specification}
\usage{
  plot_spec(plot, width = 600, height = 400,
    padding = c(20, 20, 20, 20), embed_data = FALSE,
    data_dir = NULL)
}
\arguments{
  \item{width}{width of visualisation in pixels}

  \item{height}{height of visualisation in pixels}

  \item{padding}{a numeric vector of length 4 giving the
  padding on the top, right, bottom and left sides
  respectively.}
}
\description{
  Output plot specification
}

