\name{vg2png}
\alias{vg2png}
\title{Render a vega plot spec to png.}
\usage{
  vg2png(src, dest = NULL)
}
\arguments{
  \item{src}{path to vega json file}

  \item{dest}{path to write output png. If not specified,
  defaults the same path as \code{src} but with extension
  \file{.png}.}
}
\description{
  Render a vega plot spec to png.
}

